refactor: move types & update Settings
