+++ src/model/CardModal.ts
@@ -1,0 +1,10 @@
+interface IUserDetails {
+  login: string;
+  avatar_url: string;
+  followers: number;
+  following: number;
+  public_repos: number;
+  html_url: string;
+}
+
+export default IUserDetails;
