+++ src/model/Limit.ts
@@ -1,0 +1,6 @@
+interface ILimitProps {
+  limit: number;
+  onLimitChange: (newLimit: number) => void;
+}
+
+export default ILimitProps;
