+++ src/model/Pagination.ts
@@ -1,0 +1,8 @@
+interface IPaginationProps {
+  totalItems: number;
+  limit: number;
+  currentPage: number;
+  onPageChange: (page: number) => void;
+}
+
+export default IPaginationProps;
