+++ src/model/Settings.ts
@@ -1,0 +1,5 @@
+interface ISettings {
+  parentClass?: string;
+}
+
+export default ISettings;
